src/EdgeModules/VideoCapture/requirements.txt
azure-iot-device~=2.7.0
opencv-python~=4.7.0.72
numpy~=1.21.4